APIでDBからユーザーIDを取得する処理をヘッダーコンポーネントに移動して共通化
