+++ src/components/AccountManageComponent.vue
@@ -1,5 +1,5 @@
 <template lang="">
-    <HeaderComponent :user="user_id"></HeaderComponent> 
+    <HeaderComponent @getUserInfo="getUserInfo"></HeaderComponent> 
     <main class="account-manage">
         <div class="title-area">
             <h2>{{ currentPath === 'register' ? 'ユーザー新規登録' : 'ユーザーログイン' }}</h2>
@@ -36,7 +36,7 @@
     <router-view></router-view>
 </template>
 <script lang="ts">
-import { ref, computed, watchEffect, onMounted } from 'vue'
+import { ref, computed, watchEffect } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import axios from 'axios'
 import registerPath from '@/assets/ts/registerPath'
@@ -116,16 +116,7 @@ export default {
 
         // ログインユーザーIDを取得
         const user_id = ref<string>('')
-            const getUserInfo = async() => {
-            const url = registerPath + 'api/manageAccount.php'
-            axios.post(url, {
-                method: 'getUserData'
-            })
-                .then(response => user_id.value = response.data.user_id)
-                .catch(error => console.log(error))
-        }
-
-        onMounted(() => getUserInfo())
+        const getUserInfo = (userId: string) => user_id.value = userId;
 
         return {
             currentURL,
@@ -136,6 +127,7 @@ export default {
             regex,
 
             submitAccountData,
+            getUserInfo,
         }
     }
 }
